+++ MemoryInspector/MemoryInspectorPascalSFloatingPoint.m
@@ -118,11 +118,10 @@
 	if (! f)
 		return FALSE;
 	*value = [NSArray arrayWithObjects:
-		[NSNumber numberWithInt:077770000 | [f exponent]],
-		[NSNumber numberWithInt:
-			077770000 | ([f negative] ? 04000 : 0) | (unsigned long) ([f mantissa] >> 53)],
-		[NSNumber numberWithInt:077770000 | (unsigned long) (([f mantissa] >> 41) & 07777)],
-		[NSNumber numberWithInt:077770000 | (unsigned long) (([f mantissa] >> 29) & 07777)],
+		[NSNumber numberWithInt:(int)(077770000 | [f exponent])],
+		[NSNumber numberWithInt:(int)(077770000 | ([f negative] ? 04000 : 0) | (unsigned long) ([f mantissa] >> 53))],
+		[NSNumber numberWithInt:(int)(077770000 | (unsigned long) (([f mantissa] >> 41) & 07777))],
+		[NSNumber numberWithInt:(int)(077770000 | (unsigned long) (([f mantissa] >> 29) & 07777))],
 		nil];
 	return TRUE;
 }
